requirements.txt
boto3==1.26.35
celery==5.2.7
Flask==2.2.2
Flask-Cors==3.0.10
python-dotenv==0.21.0
redis==4.4.0
requests==2.28.1
wheel
gunicorn
git+https://github.com/openai/whisper.git

